:bug: Fix terminal's "Send" button in OctoPrint >= 1.3.5
To support full width input-with-add-on stuff I had to change the
send <button> to a send <a class="btn"> in OctoPrint (buttons
apparently being enforced as display: inline-block by the browser
didn't sit well with the Bootstrap 2 workaround for input-block-level
to work with add-on inputs). This made the styles in this plugin not
properly apply anymore.

This patch should fix that and hopefully also be fully backwards
compatible - better test the latter though :)

Signed-off-by: Gina Häußge <gina@octoprint.org>